Updated CSS styles, removed images, and changed navbar color scheme
The commit includes changes to the CSS file (style.css) with the addition of a new background color for the body and changes to the navbar color scheme. Six images have been deleted from the 'img/portfolio' and 'img/backgound' directories. The 'bg-light' class in the navbar has been changed to 'bg-dark-weak'.